# platform/data_api/routes/rankings.py
# -------------------------------------------------------------------------------------------------
#  Copyright (C) 2015-2026 Nautech Systems Pty Ltd. All rights reserved.
#  https://nautechsystems.io
#
#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
#  You may not use this file except in compliance with the License.
#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
#
#  Unless required by applicable law or agreed to in writing, software
#  distributed under the License is distributed on an "AS IS" BASIS,
#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#  See the License for the specific language governing permissions and
#  limitations under the License.
# -------------------------------------------------------------------------------------------------
"""
`GET /api/rankings` -- verbatim passthrough of `ranking_engine`'s cached `rankings:live`
message (AD-F2: never recompute rank/volatility), with exactly one sanctioned reshape:
`ranks` -> `items` (epics AC1's literal `{"items": [...], "updated_at": ...}` shape).
Every other field (`mode`, `stale_instrument_ids`, and every field inside each ranking
entry) passes through byte-for-byte from the cached Redis payload.

The Technicals tab's routes are format + transport only (Story 24.2): the column selection is
`views.preferences` (`screener_columns.toml`), each coin's values
`views.ranking_columns.technicals_values` (the chart's own indicator dispatch over the chart's own
candles). This module validates the untrusted request, passes its own `CATALOG_PATH`/
`CANDLES_DB_DIR` in, keeps the short TTL cache and maps failures (bad params -> 400, one coin's
failed read -> that coin's `errors` entry).

`PUT /api/rankings/mode` (Story 25.1a) is the one ranking-mode switch now that rankings are
web-only: it publishes the message `ranking_engine` reads on `ranking:control`, byte-identical to
what the retired TUI `m` key sent, and never touches the mode itself -- the page shows the new
mode only once `rankings:live` carries it (publish-and-wait, last write wins).

`GET`/`PUT /api/rankings/filter-presets` (Story 33.7) are the page's named filter presets, one
server-side list (`views.preferences`' `screener_filter_presets.toml`), so a preset follows the
operator to another browser.
"""

import json
import logging
import time
import tomllib
from collections.abc import Sequence
from pathlib import Path
from typing import Any
from typing import Literal

import redis.asyncio as aioredis
from fastapi import APIRouter
from fastapi import HTTPException
from fastapi import Request
from fastapi.concurrency import run_in_threadpool
from observability import error_ledger
from pydantic import BaseModel
from pydantic import ConfigDict
from views import indicator_picker
from views import preferences
from views import ranking_columns

from data_api import buses
from data_api import settings
from data_api.routes import indicators as _indicators
from data_api.settings import CANDLES_DB_DIR
from data_api.settings import CATALOG_PATH
from data_api.settings import REDIS_URL
from data_api.settings import SCREENER_COLUMNS_CONFIG_PATH


router = APIRouter()
logger = logging.getLogger(__name__)


class RankingsResponse(BaseModel):
    items: list[dict]
    updated_at: int
    mode: str
    stale_instrument_ids: list[str]


@router.get("/api/rankings")
def get_rankings() -> RankingsResponse:
    """
    503 before the bus has cached its first message -- an honest transient state, not a
    fabricated empty snapshot (I/O matrix). Referencing `buses.bus` via the module
    (not importing the `bus` name directly) so tests can `monkeypatch.setattr(buses,
    "bus", RankingsBus())` to exercise this route against an isolated cache.
    """
    latest = buses.bus.latest
    if latest is None:
        raise HTTPException(status_code=503, detail="Rankings not yet available")
    return RankingsResponse(
        items=latest["ranks"],
        updated_at=latest["updated_at"],
        mode=latest["mode"],
        stale_instrument_ids=latest.get("stale_instrument_ids", []),
    )


# ---------------------------------------------------------------------------------------------
# Story 25.1a: ranking-mode switch
# ---------------------------------------------------------------------------------------------

# `ranking_engine`'s control channel (a frozen published name, epic 25). Module-level so tests
# can monkeypatch a unique channel: the local stack's live `ranking_engine` listens on the real
# one, and a test publish there would flip the live mode.
RANKING_CONTROL_CHANNEL = "ranking:control"

# A blackholed Redis must fail the request promptly (503), not hang it -- and with it the page's
# mode buttons, which stay disabled while a switch is in flight.
_REDIS_TIMEOUT_SECONDS = 2.0


class RankingModeRequest(BaseModel):
    """The whole request: one of the two modes `ranking_engine` accepts, nothing else."""

    model_config = ConfigDict(extra="forbid")

    mode: Literal["volume", "volatility"]


class RankingModeResponse(BaseModel):
    mode: Literal["volume", "volatility"]


@router.put(
    "/api/rankings/mode",
    status_code=202,
    responses={503: {"description": "No subscriber on `ranking:control`, or Redis failed"}},
)
async def put_ranking_mode(body: RankingModeRequest) -> RankingModeResponse:
    """
    Ask `ranking_engine` to switch the global ranking mode. 202, not 200: the switch is
    applied asynchronously, and `rankings:live` confirms it.

    The payload is `json.dumps({"mode": mode})`, the exact bytes the retired TUI toggle
    published (replay test: `tests/test_rankings_mode.py`). A short-lived connection per call,
    as the TUI had: a mode switch is a rare human action, not worth a held publisher.

    503 when nobody received it (`publish()` counts 0 subscribers, i.e. `ranking_engine` is
    down) or Redis fails or stalls past `_REDIS_TIMEOUT_SECONDS`. The TUI ignored both, so a
    switch sent while the engine was down vanished silently; here the caller sees it (DATA-07).

    Known limit: `publish()` counts every subscriber of the channel, so a stray `redis-cli
    SUBSCRIBE ranking:control` makes a 202 while `ranking_engine` is down. The page still never
    shows the new mode until `rankings:live` carries it. Upgrade path: an acknowledged command
    (the engine echoes a request id on `rankings:live`) once `RankingBoard.switch_mode` exists
    (Story 25.2).
    """
    payload = json.dumps({"mode": body.mode})
    try:
        async with aioredis.Redis.from_url(
            REDIS_URL,
            decode_responses=True,
            socket_connect_timeout=_REDIS_TIMEOUT_SECONDS,
            socket_timeout=_REDIS_TIMEOUT_SECONDS,
        ) as client:
            receivers = await client.publish(RANKING_CONTROL_CHANNEL, payload)
    except Exception as exc:
        raise HTTPException(
            status_code=503, detail=f"failed to publish to {RANKING_CONTROL_CHANNEL}: {exc!r}"
        ) from exc
    if receivers == 0:
        raise HTTPException(
            status_code=503,
            detail=f"no ranking_engine subscribed to {RANKING_CONTROL_CHANNEL}",
        )
    return RankingModeResponse(mode=body.mode)


# ---------------------------------------------------------------------------------------------
# Story 17.5: Technicals tab -- screener-wide column selection + bulk per-instrument values
# ---------------------------------------------------------------------------------------------

# The column timeframes the UI offers; anything else is a 400, not a silent fallback.
_TECHNICALS_BAR_SIZES = (60, 300, 900, 3600, 14400, 86400)
# Bulk values are identical for every viewer of the same entries/ranked set -- a short TTL cache
# keeps a second tab or a slow poll from stacking another full per-coin catalog pass.
_TECHNICALS_CACHE_TTL_S = 90.0
_technicals_cache: dict[str, tuple[float, dict[str, dict[str, float | None]]]] = {}


class TechnicalsColumn(_indicators.PickerEntry):
    """A per-coin picker entry plus the bar size it is computed on."""

    bar_seconds: int = preferences.DEFAULT_BAR_SECONDS


class TechnicalsRequestEntry(_indicators.IndicatorValueRequestEntry):
    bar_seconds: int = preferences.DEFAULT_BAR_SECONDS


@router.get("/api/rankings/technicals-columns")
def get_technicals_columns() -> list[TechnicalsColumn]:
    try:
        entries = preferences.load_screener_columns(Path(SCREENER_COLUMNS_CONFIG_PATH))
    except (tomllib.TOMLDecodeError, KeyError, TypeError) as exc:
        raise HTTPException(
            status_code=500, detail=f"screener_columns.toml is corrupt: {exc}"
        ) from exc
    except OSError as exc:
        raise HTTPException(
            status_code=500, detail=f"failed to read screener_columns.toml: {exc}"
        ) from exc
    return [TechnicalsColumn(**vars(e)) for e in entries]


def _require_valid_indicators(entries: Sequence[tuple[str, Any]]) -> None:
    """
    Refuse (400) a name the catalogs don't know or params its replay refuses: saved, either would
    make every later values poll fail (`indicators.check_entry_params`, the picker's one rule).
    """
    known = indicator_picker.merged_catalog()
    unknown = [name for name, _ in entries if name not in known]
    if unknown:
        raise HTTPException(status_code=400, detail=f"unknown indicator(s): {unknown}")
    for name, params in entries:
        _indicators.check_entry_params(name, params)


@router.put(
    "/api/rankings/technicals-columns",
    openapi_extra=_indicators.json_array_body("TechnicalsColumn"),
)
async def put_technicals_columns(request: Request) -> dict[str, bool]:
    """
    Persist the FULL screener-wide column list (add/remove/reorder/param change). The body is a
    JSON array of `TechnicalsColumn`, read raw so a malformed one is a 400, never a 422.
    """
    try:
        payload = await request.json()
        entries = [
            preferences.ColumnEntry(
                name=e["name"],
                params=e.get("params", {}),
                category=e["category"],
                bar_seconds=e.get("bar_seconds", preferences.DEFAULT_BAR_SECONDS),
            )
            for e in payload
        ]
    except (json.JSONDecodeError, KeyError, TypeError) as exc:
        raise HTTPException(status_code=400, detail=f"invalid columns payload: {exc}") from exc
    if not all(isinstance(e.params, dict) for e in entries):
        raise HTTPException(
            status_code=400, detail="invalid columns payload: params must be an object"
        )
    if len(entries) > _indicators._MAX_INDICATOR_VALUES_ENTRIES:
        raise HTTPException(
            status_code=400,
            detail=f"too many columns: {len(entries)} > {_indicators._MAX_INDICATOR_VALUES_ENTRIES}",
        )
    if any(e.bar_seconds not in _TECHNICALS_BAR_SIZES for e in entries):
        raise HTTPException(
            status_code=400, detail=f"bar_seconds must be one of {_TECHNICALS_BAR_SIZES}"
        )
    _require_valid_indicators([(e.name, e.params) for e in entries])
    try:
        preferences.save_screener_columns(entries, Path(SCREENER_COLUMNS_CONFIG_PATH))
    except OSError as exc:
        raise HTTPException(
            status_code=500, detail=f"failed to write screener_columns.toml: {exc}"
        ) from exc
    return {"ok": True}


class TechnicalsValuesResponse(BaseModel):
    # instrument_id -> "{entry_index}.{output_attr}" -> latest value (None: no data yet).
    # Keyed by the request entry's position, not the chart's `indicator_id` scheme, so the
    # client never has to reproduce that server-side param-formatting to find its column.
    values: dict[str, dict[str, float | None]]
    # instrument_id -> why its values are missing. A coin listed here has NO data because its
    # computation failed -- distinct from a coin with `None` values (genuinely no data yet).
    errors: dict[str, str] = {}


@router.get("/api/rankings/technicals-values")
def get_technicals_values(entries: str) -> TechnicalsValuesResponse:
    """
    Latest value of every requested indicator for every currently-ranked instrument.

    Computes nothing itself (AD-F2): `views.ranking_columns.technicals_values` dispatches to the
    chart's own indicator replay, so a column always equals what that coin's chart shows. Known limit: sequential per-coin catalog
    reads behind a 90s TTL cache (> the client's 60s poll, or it never hits) (one live key) -- no single-flight, add one if concurrent
    viewers ever load the box.
    """
    parsed = _indicators._parse_entries(entries, TechnicalsRequestEntry)
    _indicators._check_sources(parsed)
    if any(e.bar_seconds not in _TECHNICALS_BAR_SIZES for e in parsed):
        raise HTTPException(
            status_code=400, detail=f"bar_seconds must be one of {_TECHNICALS_BAR_SIZES}"
        )
    _require_valid_indicators([(e.name, e.params) for e in parsed])
    latest = buses.bus.latest
    if latest is None:
        raise HTTPException(status_code=503, detail="Rankings not yet available")
    iids = [row["instrument_id"] for row in latest["ranks"]]
    # Order-independent: rank order shifts constantly and must not defeat the cache.
    cache_key = json.dumps([entries, sorted(iids)])
    cached = _technicals_cache.get(cache_key)
    if cached is not None and time.monotonic() - cached[0] < _TECHNICALS_CACHE_TTL_S:
        return TechnicalsValuesResponse(values=cached[1])
    now_ns = time.time_ns()
    result: dict[str, dict[str, float | None]] = {}
    errors: dict[str, str] = {}
    for iid in iids:
        try:
            result[iid] = ranking_columns.technicals_values(
                iid, parsed, now_ns, catalog_path=CATALOG_PATH, candles_dir=CANDLES_DB_DIR
            )
        except ValueError as exc:  # bad indicator params -- the same for every coin, client input
            raise HTTPException(status_code=400, detail=f"invalid indicator entry: {exc}") from exc
        except Exception as exc:
            error_ledger.record("technicals.values", f"technicals values failed for {iid}", exc)
            errors[iid] = repr(exc)
    if not errors:  # never cache a failure: the next poll must retry it
        _technicals_cache.clear()  # one live key at a time; bounded (MEM-01)
        _technicals_cache[cache_key] = (time.monotonic(), result)
    return TechnicalsValuesResponse(values=result, errors=errors)


# ---------------------------------------------------------------------------------------------
# Story 33.7: saved filter presets
# ---------------------------------------------------------------------------------------------


class FilterPresetConditionItem(BaseModel):
    """One stored condition; its display precision is re-derived by the page on recall."""

    field: str
    op: Literal[">", "<", ">=", "<=", "="]
    value: float | str


class FilterPresetItem(BaseModel):
    name: str
    conditions: list[FilterPresetConditionItem]


class FilterPresetsResponse(BaseModel):
    presets: list[FilterPresetItem]


def _filter_presets_path() -> Path:
    # Read per call (not bound at import) so a test or an operator override is honoured.
    return Path(settings.SCREENER_FILTER_PRESETS_PATH)


def _presets_response(presets: list[preferences.FilterPreset]) -> FilterPresetsResponse:
    return FilterPresetsResponse(
        presets=[
            FilterPresetItem(
                name=p.name,
                conditions=[
                    FilterPresetConditionItem.model_validate(vars(c)) for c in p.conditions
                ],
            )
            for p in presets
        ]
    )


def _load_presets_or_500() -> list[preferences.FilterPreset]:
    try:
        return preferences.load_filter_presets(_filter_presets_path())
    except (tomllib.TOMLDecodeError, UnicodeDecodeError, preferences.FilterPresetError) as exc:
        # Fail loud (DATA-02): a hand-edited file that no longer parses is never shown as "none".
        raise HTTPException(
            status_code=500, detail=f"screener_filter_presets.toml is corrupt: {exc}"
        ) from exc
    except OSError as exc:
        raise HTTPException(
            status_code=500, detail=f"failed to read screener_filter_presets.toml: {exc}"
        ) from exc


@router.get("/api/rankings/filter-presets")
def get_filter_presets() -> FilterPresetsResponse:
    """Return every saved preset; none saved yet is `[]`, a corrupt file a 500."""
    return _presets_response(_load_presets_or_500())


def _store_presets(presets: list[preferences.FilterPreset]) -> None:
    with _indicators.PREFERENCES_LOCK:
        try:
            preferences.save_filter_presets(presets, _filter_presets_path())
        except OSError as exc:
            raise HTTPException(
                status_code=500, detail=f"failed to write screener_filter_presets.toml: {exc}"
            ) from exc


@router.put(
    "/api/rankings/filter-presets",
    openapi_extra={
        "requestBody": {
            "required": True,
            "content": {
                "application/json": {
                    "schema": {"$ref": "#/components/schemas/FilterPresetsResponse"}
                }
            },
        }
    },
)
async def put_filter_presets(request: Request) -> FilterPresetsResponse:
    """
    Replace the whole preset list with the body's `{"presets": [...]}` and return what is stored.

    Invalid JSON is a 400; a body that is not a storable preset list a 422 naming the field
    (`presets[1].conditions[0].value: ...`), and nothing is written unless all of it passes. The
    write is a full rewrite under `PREFERENCES_LOCK`, published atomically.

    Known limit: whole-list last-write-wins with no version, so two tabs or browsers saving presets
    at once overwrite each other (the later save wins), and the lock is in-process only. Upgrade
    path: a `version` returned by the GET and sent back by the PUT, a 409 on a mismatch.
    """
    try:
        payload = await request.json()
    except (json.JSONDecodeError, UnicodeDecodeError) as exc:
        raise HTTPException(status_code=400, detail=f"invalid presets payload: {exc}") from exc
    try:
        presets = preferences.validate_filter_presets(payload)
    except preferences.FilterPresetError as exc:
        raise HTTPException(status_code=422, detail=str(exc)) from exc
    await run_in_threadpool(_store_presets, presets)
    return _presets_response(presets)