# platform/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rankings/mode": {
      "put": {
        "summary": "Put Ranking Mode",
        "description": "Ask `ranking_engine` to switch the global ranking mode. 202, not 200: the switch is\napplied asynchronously, and `rankings:live` confirms it.\n\nThe payload is `json.dumps({\"mode\": mode})`, the exact bytes the retired TUI toggle\npublished (replay test: `tests/test_rankings_mode.py`). A short-lived connection per call,\nas the TUI had: a mode switch is a rare human action, not worth a held publisher.\n\n503 when nobody received it (`publish()` counts 0 subscribers, i.e. `ranking_engine` is\ndown) or Redis fails or stalls past `_REDIS_TIMEOUT_SECONDS`. The TUI ignored both, so a\nswitch sent while the engine was down vanished silently; here the caller sees it (DATA-07).\n\nKnown limit: `publish()` counts every subscriber of the channel, so a stray `redis-cli\nSUBSCRIBE ranking:control` makes a 202 while `ranking_engine` is down. The page still never\nshows the new mode until `rankings:live` carries it. Upgrade path: an acknowledged command\n(the engine echoes a request id on `rankings:live`) once `RankingBoard.switch_mode` exists\n(Story 25.2).",
        "operationId": "put_ranking_mode_api_rankings_mode_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingModeResponse"
                }
              }
            }
          },
          "503": {
            "description": "No subscriber on `ranking:control`, or Redis failed"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rankings/technicals-columns": {
      "get": {
        "summary": "Get Technicals Columns",
        "operationId": "get_technicals_columns_api_rankings_technicals_columns_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TechnicalsColumn"
                  },
                  "type": "array",
                  "title": "Response Get Technicals Columns Api Rankings Technicals Columns Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Technicals Columns",
        "description": "Persist the FULL screener-wide column list (add/remove/reorder/param change). The body is a\nJSON array of `TechnicalsColumn`, read raw so a malformed one is a 400, never a 422.",
        "operationId": "put_technicals_columns_api_rankings_technicals_columns_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TechnicalsColumn"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "type": "object",
                  "title": "Response Put Technicals Columns Api Rankings Technicals Columns Put"
                }
              }
            }
          }
        }
      }
    },
    "/api/rankings/technicals-values": {
      "get": {
        "summary": "Get Technicals Values",
        "description": "Latest value of every requested indicator for every currently-ranked instrument.\n\nComputes nothing itself (AD-F2): `views.ranking_columns.technicals_values` dispatches to the\nchart's own indicator replay, so a column always equals what that coin's chart shows. Known limit: sequential per-coin catalog\nreads behind a 90s TTL cache (> the client's 60s poll, or it never hits) (one live key) -- no single-flight, add one if concurrent\nviewers ever load the box.",
        "operationId": "get_technicals_values_api_rankings_technicals_values_get",
        "parameters": [
          {
            "name": "entries",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entries"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalsValuesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rankings/filter-presets": {
      "get": {
        "summary": "Get Filter Presets",
        "description": "Return every saved preset; none saved yet is `[]`, a corrupt file a 500.",
        "operationId": "get_filter_presets_api_rankings_filter_presets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterPresetsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Put Filter Presets",
        "description": "Replace the whole preset list with the body's `{\"presets\": [...]}` and return what is stored.\n\nInvalid JSON is a 400; a body that is not a storable preset list a 422 naming the field\n(`presets[1].conditions[0].value: ...`), and nothing is written unless all of it passes. The\nwrite is a full rewrite under `PREFERENCES_LOCK`, published atomically.\n\nKnown limit: whole-list last-write-wins with no version, so two tabs or browsers saving presets\nat once overwrite each other (the later save wins), and the lock is in-process only. Upgrade\npath: a `version` returned by the GET and sent back by the PUT, a 409 on a mismatch.",
        "operationId": "put_filter_presets_api_rankings_filter_presets_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterPresetsResponse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilterPresetsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FilterPresetConditionItem": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "op": {
            "type": "string",
            "enum": [
              ">",
              "<",
              ">=",
              "<=",
              "="
            ],
            "title": "Op"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "field",
          "op",
          "value"
        ],
        "title": "FilterPresetConditionItem",
        "description": "One stored condition; its display precision is re-derived by the page on recall."
      },
      "FilterPresetItem": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/FilterPresetConditionItem"
            },
            "type": "array",
            "title": "Conditions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "conditions"
        ],
        "title": "FilterPresetItem"
      },
      "FilterPresetsResponse": {
        "properties": {
          "presets": {
            "items": {
              "$ref": "#/components/schemas/FilterPresetItem"
            },
            "type": "array",
            "title": "Presets"
          }
        },
        "type": "object",
        "required": [
          "presets"
        ],
        "title": "FilterPresetsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankingModeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "volume",
              "volatility"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "RankingModeRequest",
        "description": "The whole request: one of the two modes `ranking_engine` accepts, nothing else."
      },
      "RankingModeResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "volume",
              "volatility"
            ],
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "RankingModeResponse"
      },
...
      "TechnicalsColumn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "bar_seconds": {
            "type": "integer",
            "title": "Bar Seconds",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "name",
          "category"
        ],
        "title": "TechnicalsColumn",
        "description": "A per-coin picker entry plus the bar size it is computed on."
      },
      "TechnicalsValuesResponse": {
        "properties": {
          "values": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "type": "object",
            "title": "Values"
          },
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Errors",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "values"
        ],
        "title": "TechnicalsValuesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}